Enabled multiple tracking requests
Due to some quirk in DrissionPage if we tried to launch multiple
instances at the same time only one would and the information it
retrieved would be returned to all objects that called any Page. This
has since been fixed and now DrissionPage will launch a new Chrome
instance for each individual request.