Tracking history updates now contain a full location object
Previously we had a simple string for the location a tracking history
update since most carriers only provide a tiny snippet of information,
but given the fact that some actually provide proper detailed
information this was changed to a full location object.

Carriers that provide little information will place that into the
addressLine field.