Added package tracking by ID and included archived flag in BaseCarrier
Users are now able to query a tracking endpoint of the API providing
only the parcel ID, allowing for outdated and archived parcels to be
requested indefinitely.

Another change that was made was the inclusion of the archived flag in
the BaseCarrier class, which now can theoretically hold every single
data point there is available about a given parcel in our system.