Started working on a script to decrypt carrier request bundles
On the website users are able to create requests for our project to
support more carriers. Since the required information needed for us to
implement a new carrier into our system can be personal/sensitive an
encrypted bundle is used to transmit this information securely and in a
privacy-presenving way.

This script will be used by us in order to decrypt the information
contained inside of these bundles.