Fixed SQL queries that included DATETIME offsets
Fixed all of the instances where unixepoch() was used and substituted
for MySQL-native DATETIME handling functions. Also refactored a bit of
the code in order to take into account the fact that the MySQL driver
returns a datetime.datetime object when querying a DATETIME column.

Due to the way MySQL does time differences we've also changed the
outdated property of parcels from seconds to days.