Started working on building a command line utility framework
In order to implement a proper management interface to the system we are
building a framework to wrap around the scripts and command-line
utilities that we've written so far.

Currently the framework is only capable of knowing the command actions
and its arguments, but isn't capable of executing any of them.